pyproject.toml
[tool.poetry]
name = "djangoauthtgbot"
version = "0.1.0"
description = ""
authors = ["BiqRedMac <redshine.nik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
python-telegram-bot = "^21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
